ContextTrackingExecutor: wrap executor instead of extending
As AbstractExecutorService (super of ThreadPoolExecutor) implements task
scheduling methods methods (ie, submit(...), invoke*(...)) using
execute(Runnable), ContextTrackingExecutor was unnecessarily calling
getActiveContexts(...) + executeWithinAll(...) 2 times for all these
methods.

The other way to solve this is to override only execute(...) and leave
submit(...) and invoke*(...) unchanged. This however relies on
implementation details of AbstractExecutorService.
Additionally, wrapping an executor allows to back
ContextTrackingExecutor with custom Executors apart from
ThreadPoolExecutor: a constructor accepting an executor to wrap was
added.

As this is a major change, version was upgraded to 2.x