InjectionContext: serialization: minimize chances for a rare Jetty race condition
Sometimes Jetty finalizes HTTP requests loooong after the response was sent.
During finalization of a request, Jetty may activate its session. If the
server is stopped after the response was sent but before the request was
finalized, it may happen that the same session is being activated and
passivated concurrently by 2 different threads. This may cause NPE if the
activating thread sets serializableScopedObjectEntries to null in
restoreAfterDeserialization(), while the passivating thread is in the middle
of storing objects to it in prepareForSerialization().